Tried 4 different ways to dynamically load the model providers.  But ended up back here.  It's working but there is a lot more tidy up.
