+++ custom_components/lednetwf_ble/manifest.json
@@ -21,7 +21,7 @@
 	"requirements": [
 		"bleak-retry-connector>=1.17.1",
 		"bleak>=0.17.0",
-                "bluetooth-sensor-state-data>=1.5.0"
+        "bluetooth-sensor-state-data>=1.5.0"
 	],
 	"version": "0.0.13"
 }
