+++ custom_components/lednetwf_ble/models/model_0x56.py
@@ -1,5 +1,5 @@
 from .model_abstractions import DefaultModelAbstraction
-from . import const
+from .. import const
 
 import logging
 LOGGER = logging.getLogger(__name__)
@@ -9,6 +9,8 @@ from homeassistant.components.light import (
     ColorMode,
     EFFECT_OFF
 )
+
+SUPPORTED_MODELS = [0x56]
 
 # 0x56 Effect data
 EFFECT_MAP_0x56 = {}
